Signal a special condition for unknown commands
To help applications using this library to handle the quite possible and
common scenario of running into an unknown command name, signal a
specific condition which can easily be handled separately. This prevents
applications from having to handle and inspect all exceptions just to
gracefully exit after this situation occurs.